pyproject.toml
[project]
name = "amem"
version = "0.1.0"
description = "An agentic memory system for AI applications, providing semantic search and memory management capabilities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core dependencies
    "chromadb>=0.4.22",
    "rank-bm25>=0.2.2",
    "nltk>=3.8.1",
    "litellm>=1.68.2",
    "numpy>=1.24.3",
    "scikit-learn>=1.3.2",
    "openai>=1.3.7",
    
    # Vector database and cloud services
    "qdrant-client>=1.7.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    
    # AWS dependencies
    "boto3>=1.38.13"
]

[tool.setuptools]
packages = ["amem"]
